requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask
# gunicorn==20.1.0
# Flask==2.0.3
requests==2.25.1
xmltodict
python-dotenv